為避免一次更動過多資料 內容先退回 Rev1486 西堤大人Rev1487的修改部份將慢慢逐次補回. 中間(1486-1499)諸位有送的修正煩請抽空再次補充&造成大家的不便尚乞海涵 謝謝大家的辛勞.
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@1500 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/net/l1j/server/serverpackets/S_SummonPack.java
@@ -25,14 +25,14 @@ import net.l1j.server.model.instance.L1SummonInstance;
 public class S_SummonPack extends ServerBasePacket {
 	private static final String S_SUMMON_PACK = "[S] S_SummonPack";
 
-	private static final short STATUS_POISON = 1;
-	private static final short STATUS_INVISIBLE = 2;
-	private static final short STATUS_PC = 4;
-	private static final short STATUS_FREEZE = 8;
-	private static final short STATUS_BRAVE = 16;
-	private static final short STATUS_ELFBRAVE = 32;
-	private static final short STATUS_FASTMOVABLE = 64;
-	private static final short STATUS_GHOST = 128;
+	private static final int STATUS_POISON = 1;
+	private static final int STATUS_INVISIBLE = 2;
+	private static final int STATUS_PC = 4;
+	private static final int STATUS_FREEZE = 8;
+	private static final int STATUS_BRAVE = 16;
+	private static final int STATUS_ELFBRAVE = 32;
+	private static final int STATUS_FASTMOVABLE = 64;
+	private static final int STATUS_GHOST = 128;
 
 	private byte[] _byte = null;
